Add files via upload
This test collection consists of two types of paths, which we refer to as narrow-ranged paths versus wide-ranged paths. The narrow-ranged paths consist of categories that are observed immediately one after the other in the Wikipedia category hierarchy. In contrast, in the wide-ranged paths subsequent categories in each path are guaranteed to have a distance of at least one hop from each other.